requirements.txt
# Core dependencies
Flask==3.0.3
Werkzeug==3.1.3
pdfplumber==0.11.4
openai==1.54.3
qdrant-client==1.12.1
python-dotenv==1.0.0
flask-talisman==1.1.0

# AI and ML dependencies
sentence-transformers==3.3.0
torch==2.5.1
transformers==4.46.2
huggingface-hub==0.26.2

# ML and Data Processing
numpy==2.0.2
scikit-learn==1.5.2
scipy==1.13.1

# Utilities
pydantic==2.9.2
tqdm==4.67.0
requests==2.32.3
python-jose==3.3.0
cryptography==43.0.3

# Testing
pytest==8.0.0

# Optional dependencies
tokenizers==0.20.3
filelock==3.16.1
packaging==24.2
pydantic_core==2.23.4
typing_extensions==4.12.2